PR feedback from Eric
