+++ packages/react-core/src/components/Drawer/examples/Drawer.md
@@ -21,92 +21,92 @@ section: components
 
 ### Basic
 
-```ts file="./BasicDrawer.tsx"
+```ts file="./DrawerBasic.tsx"
 ```
 
 ### Panel on right
 
-```ts file="./PanelRightDrawer.tsx"
+```ts file="./DrawerPanelRight.tsx"
 ```
 
 ### Panel on left
 
-```ts file="./PanelLeftDrawer.tsx"
+```ts file="./DrawerPanelLeft.tsx"
 ```
 
 ### Panel on bottom
 
-```ts file="./PanelBottomDrawer.tsx"
+```ts file="./DrawerPanelBottom.tsx"
 ```
 
 ### Basic inline
 
-```ts file="./BasicInlineDrawer.tsx"
+```ts file="./DrawerBasicInline.tsx"
 ```
 
 ### Inline panel on right
 
-```ts file="./InlinePanelRightDrawer.tsx"
+```ts file="./DrawerInlinePanelRight.tsx"
 ```
 
 ### Inline panel on left
 
-```ts file="./InlinePanelLeftDrawer.tsx"
+```ts file="./DrawerInlinePanelLeft.tsx"
 ```
 
 ### Stacked content body elements
 
-```ts file="./StackedContentBodyElementsDrawer.tsx"
+```ts file="./DrawerStackedContentBodyElements.tsx"
 ```
 
 ### Modified content padding
 
-```ts file="ModifiedContentPaddingDrawer.tsx"
+```ts file="DrawerModifiedContentPadding.tsx"
 ```
 
 ### Modified panel padding
 
-```ts file="ModifiedPanelPaddingDrawer.tsx"
+```ts file="DrawerModifiedPanelPadding.tsx"
 ```
 
 ### Additional section above drawer content
 
-```ts file="AdditionalSectionAboveContentDrawer.tsx"
+```ts file="DrawerAdditionalSectionAboveContent.tsx"
 ```
 
 ### Static drawer
 
 **Note:** For mobile viewports, all drawer variants behave the same way. At the `md` breakpoint, or where `.pf-m-static{-on-[lg, xl, 2xl]}` is applied, the `static drawer` variant’s `close button` is automatically hidden because the drawer panel doesn’t close by design.
 
-```ts file="StaticDrawer.tsx"
+```ts file="DrawerStatic.tsx"
 ```
 
 ### Breakpoint
 
-```ts file="BreakpointDrawer.tsx"
+```ts file="DrawerBreakpoint.tsx"
 ```
 
 ### Resizable on right
 
-```ts file="ResizableOnRightDrawer.tsx"
+```ts file="DrawerResizableOnRight.tsx"
 ```
 
 ### Resizable on left
 
-```ts file="ResizableOnLeftDrawer.tsx"
+```ts file="DrawerResizableOnLeft.tsx"
 ```
 
 ### Resizable on bottom
 
-```ts file="ResizableOnBottomDrawer.tsx"
+```ts file="DrawerResizableOnBottom.tsx"
 ```
 
 ### Resizable on inline
 
-```ts file="ResizableOnInlineDrawer.tsx"
+```ts file="DrawerResizableOnInline.tsx"
 ```
 
 ### Panel with light-200 background
 
-```ts file="LightGrayDrawer.tsx"
+```ts file="DrawerLightGray.tsx"
 ```
