+++ packages/react-core/src/components/Drawer/examples/DrawerLightGray.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import {
+  Checkbox,
   Drawer,
   DrawerPanelContent,
   DrawerContent,
@@ -7,11 +8,17 @@ import {
   DrawerHead,
   DrawerActions,
   DrawerCloseButton,
-  Button
+  DrawerSection,
+  Button,
+  DrawerColorVariant
 } from '@patternfly/react-core';
 
-export const ResizableOnRightDrawer: React.FunctionComponent = () => {
+export const DrawerLightGray: React.FunctionComponent = () => {
   const [isExpanded, setIsExpanded] = React.useState(false);
+  const [panelGray, setPanelGray] = React.useState(true);
+  const [contentGray, setContentGray] = React.useState(false);
+  const [sectionGray, setSectionGray] = React.useState(false);
+
   const drawerRef = React.useRef<HTMLDivElement>();
 
   const onExpand = () => {
@@ -26,8 +33,20 @@ export const ResizableOnRightDrawer: React.FunctionComponent = () => {
     setIsExpanded(false);
   };
 
+  const togglePanelGray = (checked: boolean) => {
+    setPanelGray(checked);
+  };
+
+  const toggleSectionGray = (checked: boolean) => {
+    setSectionGray(checked);
+  };
+
+  const toggleContentGray = (checked: boolean) => {
+    setContentGray(checked);
+  };
+
   const panelContent = (
-    <DrawerPanelContent isResizable defaultSize={'500px'} minSize={'200px'}>
+    <DrawerPanelContent colorVariant={panelGray ? DrawerColorVariant.light200 : DrawerColorVariant.default}>
       <DrawerHead>
         <span tabIndex={isExpanded ? 0 : -1} ref={drawerRef}>
           drawer-panel
@@ -44,11 +63,42 @@ export const ResizableOnRightDrawer: React.FunctionComponent = () => {
 
   return (
     <React.Fragment>
+      <Checkbox
+        label="Gray panel"
+        isChecked={panelGray}
+        onChange={togglePanelGray}
+        aria-label="Gray panel checkbox"
+        id="toggle-gray-panel"
+        name="toggle-gray-panel"
+      />
+      <Checkbox
+        label="Gray content"
+        isChecked={contentGray}
+        onChange={toggleContentGray}
+        aria-label="Gray content checkbox"
+        id="toggle-gray-content"
+        name="toggle-gray-content"
+      />
+      <Checkbox
+        label="Gray section"
+        isChecked={sectionGray}
+        onChange={toggleSectionGray}
+        aria-label="Gray section checkbox"
+        id="toggle-gray-section"
+        name="toggle-gray-section"
+      />
+      <br />
       <Button aria-expanded={isExpanded} onClick={onClick}>
         Toggle drawer
       </Button>
-      <Drawer isExpanded={isExpanded} onExpand={onExpand} position="left">
-        <DrawerContent panelContent={panelContent}>
+      <Drawer isExpanded={isExpanded} onExpand={onExpand}>
+        <DrawerSection colorVariant={sectionGray ? DrawerColorVariant.light200 : DrawerColorVariant.default}>
+          drawer-section
+        </DrawerSection>
+        <DrawerContent
+          colorVariant={contentGray ? DrawerColorVariant.light200 : DrawerColorVariant.default}
+          panelContent={panelContent}
+        >
           <DrawerContentBody>{drawerContent}</DrawerContentBody>
         </DrawerContent>
       </Drawer>
