+++ packages/react-core/src/components/Drawer/examples/DrawerModifiedPanelPadding.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  Checkbox,
   Drawer,
   DrawerPanelContent,
   DrawerContent,
@@ -8,17 +7,11 @@ import {
   DrawerHead,
   DrawerActions,
   DrawerCloseButton,
-  DrawerSection,
-  Button,
-  DrawerColorVariant
+  Button
 } from '@patternfly/react-core';
 
-export const LightGrayDrawer: React.FunctionComponent = () => {
+export const DrawerModifiedPanelPadding: React.FunctionComponent = () => {
   const [isExpanded, setIsExpanded] = React.useState(false);
-  const [panelGray, setPanelGray] = React.useState(true);
-  const [contentGray, setContentGray] = React.useState(false);
-  const [sectionGray, setSectionGray] = React.useState(false);
-
   const drawerRef = React.useRef<HTMLDivElement>();
 
   const onExpand = () => {
@@ -33,21 +26,9 @@ export const LightGrayDrawer: React.FunctionComponent = () => {
     setIsExpanded(false);
   };
 
-  const togglePanelGray = (checked: boolean) => {
-    setPanelGray(checked);
-  };
-
-  const toggleSectionGray = (checked: boolean) => {
-    setSectionGray(checked);
-  };
-
-  const toggleContentGray = (checked: boolean) => {
-    setContentGray(checked);
-  };
-
   const panelContent = (
-    <DrawerPanelContent colorVariant={panelGray ? DrawerColorVariant.light200 : DrawerColorVariant.default}>
-      <DrawerHead>
+    <DrawerPanelContent>
+      <DrawerHead hasNoPadding>
         <span tabIndex={isExpanded ? 0 : -1} ref={drawerRef}>
           drawer-panel
         </span>
@@ -63,42 +44,11 @@ export const LightGrayDrawer: React.FunctionComponent = () => {
 
   return (
     <React.Fragment>
-      <Checkbox
-        label="Gray panel"
-        isChecked={panelGray}
-        onChange={togglePanelGray}
-        aria-label="Gray panel checkbox"
-        id="toggle-gray-panel"
-        name="toggle-gray-panel"
-      />
-      <Checkbox
-        label="Gray content"
-        isChecked={contentGray}
-        onChange={toggleContentGray}
-        aria-label="Gray content checkbox"
-        id="toggle-gray-content"
-        name="toggle-gray-content"
-      />
-      <Checkbox
-        label="Gray section"
-        isChecked={sectionGray}
-        onChange={toggleSectionGray}
-        aria-label="Gray section checkbox"
-        id="toggle-gray-section"
-        name="toggle-gray-section"
-      />
-      <br />
       <Button aria-expanded={isExpanded} onClick={onClick}>
         Toggle drawer
       </Button>
       <Drawer isExpanded={isExpanded} onExpand={onExpand}>
-        <DrawerSection colorVariant={sectionGray ? DrawerColorVariant.light200 : DrawerColorVariant.default}>
-          drawer-section
-        </DrawerSection>
-        <DrawerContent
-          colorVariant={contentGray ? DrawerColorVariant.light200 : DrawerColorVariant.default}
-          panelContent={panelContent}
-        >
+        <DrawerContent panelContent={panelContent}>
           <DrawerContentBody>{drawerContent}</DrawerContentBody>
         </DrawerContent>
       </Drawer>
