+++ packages/react-core/src/components/Drawer/examples/DrawerResizableOnBottom.tsx
@@ -2,15 +2,15 @@ import React from 'react';
 import {
   Drawer,
   DrawerPanelContent,
-  DrawerContent,
-  DrawerContentBody,
   DrawerHead,
   DrawerActions,
   DrawerCloseButton,
-  Button
+  Button,
+  DrawerContent,
+  DrawerContentBody
 } from '@patternfly/react-core';
 
-export const PanelRightDrawer: React.FunctionComponent = () => {
+export const DrawerResizableOnBottom: React.FunctionComponent = () => {
   const [isExpanded, setIsExpanded] = React.useState(false);
   const drawerRef = React.useRef<HTMLDivElement>();
 
@@ -25,9 +25,8 @@ export const PanelRightDrawer: React.FunctionComponent = () => {
   const onCloseClick = () => {
     setIsExpanded(false);
   };
-
   const panelContent = (
-    <DrawerPanelContent>
+    <DrawerPanelContent isResizable defaultSize={'200px'} minSize={'100px'}>
       <DrawerHead>
         <span tabIndex={isExpanded ? 0 : -1} ref={drawerRef}>
           drawer-panel
@@ -47,11 +46,13 @@ export const PanelRightDrawer: React.FunctionComponent = () => {
       <Button aria-expanded={isExpanded} onClick={onClick}>
         Toggle drawer
       </Button>
-      <Drawer isExpanded={isExpanded} position="right" onExpand={onExpand}>
-        <DrawerContent panelContent={panelContent}>
-          <DrawerContentBody>{drawerContent}</DrawerContentBody>
-        </DrawerContent>
-      </Drawer>
+      <div style={{ height: '400px' }}>
+        <Drawer isExpanded={isExpanded} onExpand={onExpand} position="bottom">
+          <DrawerContent panelContent={panelContent}>
+            <DrawerContentBody>{drawerContent}</DrawerContentBody>
+          </DrawerContent>
+        </Drawer>
+      </div>
     </React.Fragment>
   );
 };
